7.0/requirements.txt
#setuptools = 2.2
babel==0.9.6
reportlab==2.5
Pillow==1.7.1
pywebdav==0.9.4.1
pyyaml==3.10
werkzeug==0.8.3
zsi==2.0-rc3
feedparser==5.1.1
gdata==2.0.16
lxml==2.3.3
psycopg2==2.7.3.2
pydot==1.0.28
pyparsing==1.5.6
python-dateutil==1.5
python-ldap==2.4.9
python-openid==2.2.5
pytz==2012b
vatnumber==1.0
vobject>=0.8.1c
xlwt==0.7.3
xlsxwriter==0.5.5
